+++ lib/olefs/objects/new_line_item.rb
@@ -25,6 +25,9 @@ module OLE_QA::Framework::OLEFS
     #   one new line item on an OLEFS PURAP E-Document.
     def set_elements
       super
+      element(:new_bib)                             {b.radio(:name => 'AttachBib', :index => 0)}
+      element(:existing_bib)                        {b.radio(:name => 'AttachBib', :index => 1)}
+      element(:bib_search_button)                   {b.input(:id => 'bibSelectExistingItemButton')}
       element(:new_bib_button)                      {b.input(:id => "bibCreateCurrentItemButton")}
       element(:item_type_selector)                  {b.select_list(:id => "newPurchasingItemLine.itemTypeDescription")}
       element(:copies_field)                        {b.text_field(:id => "newPurchasingItemLine.oleItemQuantity")}